// types/dataTypes/userdataType.ts
export type UserDataType = {
  user_id: number;
  user_name: string;
  user_email: string;
  user_gender: "male" | "female";
  school_id: number;
  school_name: string;
  school_desc: string;
  school_address: string;
  school_created: string;
  user_role: number;
  role_name: string;
  user_created: string;
};